python--jira-query/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.3
jira==3.2.0
keyring==23.5.0
kiwisolver==1.4.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
